proof of concept for cleaning up forms
