+++ frontend/src/ui/home/Home.js
@@ -1,7 +1,7 @@
 import React, {useEffect} from 'react';
 import {useSelector, useDispatch} from "react-redux";
 import {TweetCard} from "../shared/components/tweetCard/TweetCard";
-import {TweetForm} from "../shared/components/tweetForm/TweetForm";
+import {TweetForm} from "./TweetForm";
 import {Container} from "react-bootstrap"
 import {fetchAllTweets} from "../../store/tweets";
